[backport#11577] Fix warnings (-Wsign-compare) when building with DEBUG_ADDRMAN
Summary:
6eddd43 Fix warnings when building with DEBUG_ADDRMAN (practicalswift)

Pull request description:

  Fix warnings when building with `DEBUG_ADDRMAN`.

  Warnings prior to this commit:

  ```
  addrman.cpp:390:24: warning: comparison of integers of different signs: 'size_type' (aka 'unsigned long') and 'int' [-Wsign-compare]
      if (vRandom.size() != nTried + nNew)
          ~~~~~~~~~~~~~~ ^  ~~~~~~~~~~~~~
  addrman.cpp:411:52: warning: comparison of integers of different signs: 'int' and 'size_type' (aka 'unsigned long') [-Wsign-compare]
          if (info.nRandomPos < 0 || info.nRandomPos >= vRandom.size() || vRandom[info.nRandomPos] != n)
                                     ~~~~~~~~~~~~~~~ ^  ~~~~~~~~~~~~~~
  addrman.cpp:419:25: warning: comparison of integers of different signs: 'size_type' (aka 'unsigned long') and 'int' [-Wsign-compare]
      if (setTried.size() != nTried)
          ~~~~~~~~~~~~~~~ ^  ~~~~~~
  addrman.cpp:421:23: warning: comparison of integers of different signs: 'size_type' (aka 'unsigned long') and 'int' [-Wsign-compare]
      if (mapNew.size() != nNew)
          ~~~~~~~~~~~~~ ^  ~~~~
  4 warnings generated.
  ```

---

Backport of Core [[https://github.com/bitcoin/bitcoin/pull/11577 | PR11577]]

Test Plan:
`#define DEBUG_ADDRMAN 1` in src/addrman.h
  ninja all check

get no warnings

Reviewers: #bitcoin_abc, PiRK

Reviewed By: PiRK

Differential Revision: https://reviews.bitcoinabc.org/D8206

+++ src/addrman.cpp
@@ -472,7 +472,7 @@ int CAddrMan::Check_() {
     std::set<int> setTried;
     std::map<int, int> mapNew;
 
-    if (vRandom.size() != nTried + nNew) {
+    if (vRandom.size() != size_t(nTried + nNew)) {
         return -7;
     }
 
@@ -500,7 +500,7 @@ int CAddrMan::Check_() {
         if (mapAddr[info] != n) {
             return -5;
         }
-        if (info.nRandomPos < 0 || info.nRandomPos >= vRandom.size() ||
+        if (info.nRandomPos < 0 || size_t(info.nRandomPos) >= vRandom.size() ||
             vRandom[info.nRandomPos] != n) {
             return -14;
         }
@@ -512,10 +512,10 @@ int CAddrMan::Check_() {
         }
     }
 
-    if (setTried.size() != nTried) {
+    if (setTried.size() != size_t(nTried)) {
         return -9;
     }
-    if (mapNew.size() != nNew) {
+    if (mapNew.size() != size_t(nNew)) {
         return -10;
     }
